requirements.txt
imageio==2.16.1
matplotlib==3.4.3
numpy==1.24.2
opencv_python==4.5.5.64
Pillow==9.2.0
PyYAML==6.0
scikit_image==0.20.0
scipy==1.10.0
seaborn==0.11.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.63.0
wand==0.6.11
wandb==0.14.2
wilds==2.0.0
